requirements.txt
beautifulsoup4
selenium
pandas
python-dotenv